[VM][Qt] Fix FTBFSs.

+++ source/src/vm/scsi_hdd.cpp
@@ -136,7 +136,7 @@ void SCSI_HDD::read_buffer(int length)
 		fio->Fseek((long)position, FILEIO_SEEK_SET);
 		while(length > 0) {
 			uint8_t tmp_buffer[SCSI_BUFFER_SIZE];
-			int tmp_length = min(length, sizeof(tmp_buffer));
+			int tmp_length = min(length, (int)sizeof(tmp_buffer));
 			
 			fio->Fread(tmp_buffer, tmp_length, 1);
 			for(int i = 0; i < tmp_length; i++) {
@@ -162,7 +162,7 @@ void SCSI_HDD::write_buffer(int length)
 		fio->Fseek((long)position, FILEIO_SEEK_SET);
 		while(length > 0) {
 			uint8_t tmp_buffer[SCSI_BUFFER_SIZE];
-			int tmp_length = min(length, sizeof(tmp_buffer));
+			int tmp_length = min(length, (int)sizeof(tmp_buffer));
 			
 			for(int i = 0; i < tmp_length; i++) {
 				tmp_buffer[i] = buffer->read();
